--- backend/database/models/index.js
@@ -1,5 +1,0 @@
-import Usuario from './userSchema.js';
-
-export {
-    Usuario
-};
