--- backend/src/controllers/index.js
@@ -1,5 +1,0 @@
-import UserController from './userController.js'
-
-export {
-    UserController
-}
