--- backend/src/repositories/index.js
@@ -1,5 +1,0 @@
-import userRepository from "./userRepository.js";
-
-export {
-    userRepository
-}
