--- backend/src/validation/index.js
@@ -1,6 +1,0 @@
-import { userSchema, userUpdateSchema } from './userValidation.js';
-
-export {
-    userSchema,
-    userUpdateSchema
-};
